Move serialization/IO and UI patches to their own files
