+++ MWSE/PatchIO.h
@@ -1,0 +1,5 @@
+#pragma once
+
+namespace mwse::patch::io {
+	void install();
+}
